# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.DefaultResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
  student.StudentRequest:
    properties:
      address_detail:
        type: string
      address_region:
        type: string
      birth_date:
        description: YYYY-MM-DD
        type: string
      birth_place:
        type: string
      class_level:
        type: string
      full_name:
        type: string
      gender:
        enum:
        - L
        - P
        type: string
      major_id:
        type: string
      nis:
        type: string
      phone:
        type: string
      photo:
        type: string
    required:
    - class_level
    - full_name
    - gender
    - major_id
    - nis
    type: object
  student.StudentResponse:
    properties:
      address_detail:
        type: string
      address_region:
        type: string
      birth_date:
        description: YYYY-MM-DD
        type: string
      birth_place:
        type: string
      class_level:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      major_id:
        type: string
      nis:
        type: string
      phone:
        type: string
      photo:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Documentation for LinguaLoop E-Learning Platform.
  title: LinguaLoop API
  version: "1.0"
paths:
...
  /students:
    get:
      description: Get a list of all students
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/student.StudentResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get all students
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: Add a new student to the system and auto-generate user account
      parameters:
      - description: Student Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/student.StudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/student.StudentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Create new student
      tags:
      - Students
  /students/{id}:
    delete:
      description: Soft delete a student by ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Delete student
      tags:
      - Students
    get:
      description: Get detailed information of a student by their ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/student.StudentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get student by ID
      tags:
      - Students
    put:
      consumes:
      - application/json
      description: Update existing student information
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      - description: Student Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/student.StudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/student.StudentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update student
      tags:
      - Students
  /students/{id}/status:
    patch:
      description: Change student status and auto-sync login capabilities
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      - description: New Status (ACTIVE/GRADUATED/TRANSFER/INACTIVE)
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  $ref: '#/definitions/student.StudentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update student status
      tags:
      - Students
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"